backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
google-genai>=1.70.0
supabase>=2.28.3
ddgs==9.11.4
python-multipart==0.0.9
fpdf2>=2.7.9
aiosmtplib>=3.0.0
pandas>=2.2.0
openpyxl>=3.1.2
